_requirements.txt
alembic==0.8.4
apispec==4.0.0
APScheduler==3.7.0
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Click==7.0
colorhash==1.0.3
configparser==5.0.2
cryptography==3.2.1
Deprecated==1.2.10
Flask==1.0.2
flask-apispec==0.7.0
Flask-Bcrypt==0.7.1
Flask-Builder==0.9
Flask-Caching==1.4.0
Flask-CLI==0.4.0
Flask-Cors==3.0.6
Flask-JWT==0.3.2
Flask-JWT-Extended==3.13.1
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==2.2.1
Flask-MonitoringDashboard==3.1.0
Flask-SQLAlchemy==2.3.2
gunicorn==19.9.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.1
lxml==4.6.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.0.0
marshmallow-sqlalchemy==0.24.2
numpy==1.20.1
psutil==5.8.0
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.9.1
python-editor==1.0.4
pytz==2021.1
PyYAML==5.3.1
pyyaml-include==1.2
raven==6.10.0
requests==2.24.0
# ruamel.yaml==0.16.12
# ruamel.yaml.clib==0.2.2
scipy==1.6.1
six==1.15.0
SQLAlchemy==1.3.23
tzlocal==2.1
unicode-slugify==0.1.3
Unidecode==1.1.1
urllib3==1.25.11
waitress==1.4.4
webargs==6.1.1
Werkzeug==0.14.1
wrapt==1.12.1
